move env out of io
